requirements.txt
streamlit
pandas
plotly
scikit-learn
google-generativeai
numpy